src/main/java/uk/gov/hmcts/pdda/business/services/pdda/CourtelHelper.java
package uk.gov.hmcts.pdda.business.services.pdda;

import java.util.Arrays;

/**
 * <p>
 * Title: CourtelHelper.
 * </p>
 * <p>
 * Description:
 * </p>
 * <p>
 * Copyright: Copyright (c) 2024
 * </p>
 * <p>
 * Company: CGI
 * </p>
 * 
 * @author Luke Gittins
 * @version 1.0
 */
public class CourtelHelper {
    
    protected static final String[] VALID_LISTS = {"DL", "DLP", "FL", "WL"};
    
    public boolean isCourtelSendableDocument(String documentType) {
        return Arrays.asList(VALID_LISTS).contains(documentType);
    }
}
